Add bounds checking and bad battery testing
When reading any value, make sure it is within bounds before reporting it. The bounds are specified on the device website.
When testing for a bad battery, make sure that we see that the battery is discharging while connected to power multiple times before reporting the battery as bad. This is because the device does periodically discharge the battery a little to test it's level.