UPSPlus shutdowns on Protected Voltage
The UPS actually fully shuts down (FSD) when the battery voltage reaches the Protected Voltage. That is not how I thought it worked.
Re-work all code to take this into account.
Set empty battery voltage to same value as protected battery voltage as for NUT they are the same thing.
Store the low battery percent as a local variable.